pyproject.toml
[project]
name = "infinity-utils"
version = "0.2.0"
description = "A collection of utility functions and classes including OAuth authentication for FastAPI applications."
authors = [
      {name = "Infinity AI Lab"}
  ]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic",
    "google-cloud-firestore",
    "fastapi",
    "httpx",
    "pyjwt[crypto]",
    "cryptography",
    "pydantic-settings",
]
